requirements.txt
# === Core ML Frameworks ===
torch==2.4.0
pytorch_lightning==1.9.5
lightning-bolts==0.7.0

# === Data Handling & Utilities ===
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
pickle-mixin

# === ML & Survival Analysis ===
scikit-learn==1.5.1
scikit-survival==0.23.0

# === NLP / Transformers ===
transformers==4.46.3
